refactor: new DebouncedInput component
